Add GitHub token validation and move dedup logic to utils
Validate that miner-provided GitHub tokens actually grant access to the
claimed repo before giving dedup priority. Moved all GitHub account
dedup functions from tournament_manager to tournament/utils.
